2006-12-09  Hendrik Boom  <hboom@cvs.gnome.org>

	* Merged in HEAD from 2006-12-07.

+++ app/plug-in/gimpplugin.c
@@ -20,6 +20,8 @@
 
 #include "config.h"
 
+#define _GNU_SOURCE
+
 #include <errno.h>
 #include <signal.h>
 #include <stdlib.h>
@@ -74,8 +76,6 @@
 #include "core/gimpcontext.h"
 #include "core/gimpprogress.h"
 
-#include "pdb/gimptemporaryprocedure.h"
-
 #include "gimpenvirontable.h"
 #include "gimpinterpreterdb.h"
 #include "gimpplugin.h"
@@ -84,6 +84,7 @@
 #include "gimpplugindebug.h"
 #include "gimppluginmanager.h"
 #include "gimppluginmanager-locale-domain.h"
+#include "gimptemporaryprocedure.h"
 #include "plug-in-def.h"
 #include "plug-in-params.h"
 
@@ -228,10 +229,11 @@ gimp_plug_in_open (GimpPlugIn         *plug_in,
    */
   if ((pipe (my_read) == -1) || (pipe (my_write) == -1))
     {
-      g_message ("Unable to run plug-in \"%s\"\n(%s)\n\npipe() failed: %s",
-                 gimp_object_get_name (GIMP_OBJECT (plug_in)),
-                 gimp_filename_to_utf8 (plug_in->prog),
-                 g_strerror (errno));
+      gimp_message (plug_in->manager->gimp, NULL, GIMP_MESSAGE_ERROR,
+                    "Unable to run plug-in \"%s\"\n(%s)\n\npipe() failed: %s",
+                    gimp_object_get_name (GIMP_OBJECT (plug_in)),
+                    gimp_filename_to_utf8 (plug_in->prog),
+                    g_strerror (errno));
       return FALSE;
     }
 
@@ -342,10 +344,11 @@ gimp_plug_in_open (GimpPlugIn         *plug_in,
                        &plug_in->pid,
                        &error))
     {
-      g_message ("Unable to run plug-in \"%s\"\n(%s)\n\n%s",
-                 gimp_object_get_name (GIMP_OBJECT (plug_in)),
-                 gimp_filename_to_utf8 (plug_in->prog),
-                 error->message);
+      gimp_message (plug_in->manager->gimp, NULL, GIMP_MESSAGE_ERROR,
+                    "Unable to run plug-in \"%s\"\n(%s)\n\n%s",
+                    gimp_object_get_name (GIMP_OBJECT (plug_in)),
+                    gimp_filename_to_utf8 (plug_in->prog),
+                    error->message);
       g_error_free (error);
       goto cleanup;
     }
@@ -428,7 +431,13 @@ gimp_plug_in_close (GimpPlugIn *plug_in,
             g_print (_("Terminating plug-in: '%s'\n"),
                      gimp_filename_to_utf8 (plug_in->prog));
 
-          status = kill (plug_in->pid, SIGKILL);
+          /*  If the plug-in opened a process group, kill the group instead
+           *  of only the plug-in, so we kill the plug-in's children too
+           */
+          if (getpgid (0) != getpgid (plug_in->pid))
+            status = kill (- plug_in->pid, SIGKILL);
+          else
+            status = kill (plug_in->pid, SIGKILL);
         }
 
       /* Wait for the process to exit. This will happen
@@ -507,8 +516,9 @@ gimp_plug_in_close (GimpPlugIn *plug_in,
       GimpPlugInProcFrame *proc_frame = list->data;
 
 #ifdef GIMP_UNSTABLE
-      g_printerr ("plug_in_close: plug-in aborted before sending its "
-                  "temporary procedure return values\n");
+      g_printerr ("plug-in '%s' aborted before sending its "
+                  "temporary procedure return values\n",
+                  gimp_object_get_name (GIMP_OBJECT (plug_in)));
 #endif
 
       if (proc_frame->main_loop &&
@@ -522,8 +532,9 @@ gimp_plug_in_close (GimpPlugIn *plug_in,
       g_main_loop_is_running (plug_in->main_proc_frame.main_loop))
     {
 #ifdef GIMP_UNSTABLE
-      g_printerr ("plug_in_close: plug-in aborted before sending its "
-                  "procedure return values\n");
+      g_printerr ("plug-in '%s' aborted before sending its "
+                  "procedure return values\n",
+                  gimp_object_get_name (GIMP_OBJECT (plug_in)));
 #endif
 
       g_main_loop_quit (plug_in->main_proc_frame.main_loop);
@@ -533,8 +544,9 @@ gimp_plug_in_close (GimpPlugIn *plug_in,
       g_main_loop_is_running (plug_in->ext_main_loop))
     {
 #ifdef GIMP_UNSTABLE
-      g_printerr ("plug_in_close: extension aborted before sending its "
-                  "extension_ack message\n");
+      g_printerr ("extension '%s' aborted before sending its "
+                  "extension_ack message\n",
+                  gimp_object_get_name (GIMP_OBJECT (plug_in)));
 #endif
 
       g_main_loop_quit (plug_in->ext_main_loop);
@@ -593,12 +605,19 @@ gimp_plug_in_recv_message (GIOChannel   *channel,
     }
 
   if (! got_message)
-    g_message (_("Plug-in crashed: \"%s\"\n(%s)\n\n"
-                 "The dying plug-in may have messed up GIMP's internal state. "
-                 "You may want to save your images and restart GIMP "
-                 "to be on the safe side."),
-               gimp_object_get_name (GIMP_OBJECT (plug_in)),
-               gimp_filename_to_utf8 (plug_in->prog));
+    {
+      GimpPlugInProcFrame *frame    = gimp_plug_in_get_proc_frame (plug_in);
+      GimpProgress        *progress = frame ? frame->progress : NULL;
+
+      gimp_message (plug_in->manager->gimp, G_OBJECT (progress),
+                    GIMP_MESSAGE_ERROR,
+                    _("Plug-in crashed: \"%s\"\n(%s)\n\n"
+                      "The dying plug-in may have messed up GIMP's internal "
+                      "state. You may want to save your images and restart "
+                      "GIMP to be on the safe side."),
+                    gimp_object_get_name (GIMP_OBJECT (plug_in)),
+                    gimp_filename_to_utf8 (plug_in->prog));
+    }
 
   g_object_unref (plug_in);
 
@@ -848,14 +867,15 @@ gimp_plug_in_menu_register (GimpPlugIn  *plug_in,
 
   if (! proc)
     {
-      g_message ("Plug-in \"%s\"\n(%s)\n"
-                 "attempted to register the menu item \"%s\" "
-                 "for the procedure \"%s\".\n"
-                 "It has however not installed that procedure.  This "
-                 "is not allowed.",
-                 gimp_object_get_name (GIMP_OBJECT (plug_in)),
-                 gimp_filename_to_utf8 (plug_in->prog),
-                 menu_path, proc_name);
+      gimp_message (plug_in->manager->gimp, NULL, GIMP_MESSAGE_ERROR,
+                    "Plug-in \"%s\"\n(%s)\n"
+                    "attempted to register the menu item \"%s\" "
+                    "for the procedure \"%s\".\n"
+                    "It has however not installed that procedure.  This "
+                    "is not allowed.",
+                    gimp_object_get_name (GIMP_OBJECT (plug_in)),
+                    gimp_filename_to_utf8 (plug_in->prog),
+                    menu_path, proc_name);
 
       return FALSE;
     }
@@ -877,23 +897,25 @@ gimp_plug_in_menu_register (GimpPlugIn  *plug_in,
 
   if (! proc->menu_label)
     {
-      g_message ("Plug-in \"%s\"\n(%s)\n"
-                 "attempted to register the menu item \"%s\" "
-                 "for procedure \"%s\".\n"
-                 "The menu label given in gimp_install_procedure() "
-                 "already contained a path.  To make this work, "
-                 "pass just the menu's label to "
-                 "gimp_install_procedure().",
-                 gimp_object_get_name (GIMP_OBJECT (plug_in)),
-                 gimp_filename_to_utf8 (plug_in->prog),
-                 menu_path, proc_name);
+      gimp_message (plug_in->manager->gimp, NULL, GIMP_MESSAGE_ERROR,
+                    "Plug-in \"%s\"\n(%s)\n"
+                    "attempted to register the menu item \"%s\" "
+                    "for procedure \"%s\".\n"
+                    "The menu label given in gimp_install_procedure() "
+                    "already contained a path.  To make this work, "
+                    "pass just the menu's label to "
+                    "gimp_install_procedure().",
+                    gimp_object_get_name (GIMP_OBJECT (plug_in)),
+                    gimp_filename_to_utf8 (plug_in->prog),
+                    menu_path, proc_name);
 
       return FALSE;
     }
 
   if (! gimp_plug_in_procedure_add_menu_path (proc, menu_path, &error))
     {
-      g_message (error->message);
+      gimp_message (plug_in->manager->gimp, NULL, GIMP_MESSAGE_ERROR,
+                    "%s", error->message);
       g_clear_error (&error);
 
       return FALSE;
